Step3: Move the pilot scene and drawables into visual/
The 3D visualization neighborhood moves as one unit: the RHI widget,
the scene graph, the camera controller, the mesh frame, and the axis
gizmo, together with RDrawable and the primitives that derive from or
lean on it (boundary, feature edges, field, scalar field, segments,
normals) and the assets they render with (material, colormap, scalar
bar).  RDrawable roots the primitives, and the scene draws them, so
splitting the set would cut the tightest coupling in the package.

This is the largest step of the reorganization: 30 files move, and
their heavy cross-includes rewrite to the visual/ paths.  RTextOverlay
stays at the root for now; it belongs to the 2D canvas neighborhood
that moves next.  No logic changes.
