+++ cpp/solvcon/pilot/RManager.hpp
@@ -15,7 +15,7 @@
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
 #include <solvcon/pilot/DrawTool.hpp>
-#include <solvcon/pilot/RDomainWidget.hpp>
+#include <solvcon/pilot/visual/RDomainWidget.hpp>
 #include <solvcon/pilot/R2DWidget.hpp>
 #include <solvcon/pilot/RAction.hpp>
 #include <solvcon/pilot/console/RPythonConsoleDockWidget.hpp>
