+++ cpp/solvcon/pilot/pilot.hpp
@@ -14,7 +14,7 @@
 
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
-#include <solvcon/pilot/RDomainWidget.hpp>
+#include <solvcon/pilot/visual/RDomainWidget.hpp>
 #include <solvcon/pilot/RManager.hpp>
 #include <solvcon/pilot/console/RPythonConsoleDockWidget.hpp>
