+++ cpp/solvcon/pilot/visual/RCameraController.cpp
@@ -3,7 +3,7 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/RCameraController.hpp> // Must be the first include.
+#include <solvcon/pilot/visual/RCameraController.hpp> // Must be the first include.
 
 #include <QQuaternion>
 #include <QtMath>
