+++ cpp/solvcon/pilot/visual/RCameraController.hpp
