+++ cpp/solvcon/pilot/visual/RColormap.hpp
