+++ cpp/solvcon/pilot/visual/RMaterial.hpp
