+++ cpp/solvcon/pilot/visual/RNormals.hpp
@@ -14,7 +14,7 @@
 
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
-#include <solvcon/pilot/RDrawable.hpp>
+#include <solvcon/pilot/visual/RDrawable.hpp>
 
 #include <solvcon/solvcon.hpp>
