+++ cpp/solvcon/pilot/visual/RScalarBar.cpp
@@ -3,7 +3,7 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/RScalarBar.hpp> // Must be the first include.
+#include <solvcon/pilot/visual/RScalarBar.hpp> // Must be the first include.
 
 #include <QColor>
 #include <QFont>
