+++ cpp/solvcon/pilot/visual/RScalarBar.hpp
@@ -14,8 +14,8 @@
 
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
-#include <solvcon/pilot/RColormap.hpp>
-#include <solvcon/pilot/RMaterial.hpp>
+#include <solvcon/pilot/visual/RColormap.hpp>
+#include <solvcon/pilot/visual/RMaterial.hpp>
 
 #include <rhi/qrhi.h>
