+++ cpp/solvcon/pilot/visual/RScalarField.cpp
@@ -3,7 +3,7 @@
  * BSD 3-Clause License, see COPYING
  */
 
-#include <solvcon/pilot/RScalarField.hpp> // Must be the first include.
+#include <solvcon/pilot/visual/RScalarField.hpp> // Must be the first include.
 
 #include <limits>
 #include <stdexcept>
