+++ cpp/solvcon/pilot/visual/RScene.hpp
@@ -15,8 +15,8 @@
 
 #include <solvcon/pilot/common/common_detail.hpp> // Must be the first include.
 
-#include <solvcon/pilot/RCameraController.hpp>
-#include <solvcon/pilot/RDrawable.hpp>
+#include <solvcon/pilot/visual/RCameraController.hpp>
+#include <solvcon/pilot/visual/RDrawable.hpp>
 
 #include <rhi/qrhi.h>
